move ICreateBlockBlobRequestOptions to different file
