+++ src/types/azure-storage-additional-types.ts
